+++ src/entities/mocks/index.ts
@@ -1,0 +1,14 @@
+export const project_data = [
+  {
+    title: '프로젝트1',
+    author: '이것은 프로젝트1 입니다.',
+  },
+  {
+    title: '프로젝트2',
+    author: '이것은 프로젝트2 입니다.',
+  },
+  {
+    title: '프로젝트3',
+    author: '이것은 프로젝트3 입니다.',
+  },
+];
