+++ src/page/index.ts
@@ -1,0 +1,4 @@
+export { LoginPage } from './auth/LoginPage';
+export { SignFormPage } from './auth/SignFormPage';
+export { SignupPage } from './auth/SignupPage';
+export { MainPage } from './workspace/MainPage';
